Refactored base Component into a function that sets the prototype and wraps the constructor for new derived Components. This now correctly supports instance initialization for common properties like owner.
