+++ src/core/component2.js
